fix(qc): cast seq.GetSize() to size_t in TrimResult brace-init
Fixes WASM build failure on wasm32-emscripten where idx_t (unsigned
long long) cannot be narrowed to std::size_t (32-bit unsigned long) in
a brace-initializer. Linux/macOS/Windows builds were unaffected
because size_t is 64-bit there.

+++ src/qc_functions.cpp
@@ -471,7 +471,7 @@ static void TrimAdaptersExecuteImpl(DataChunk &args, Vector &result, bool adapte
 		}
 
 		// Leftmost trim_start across all candidates wins.
-		miint::qc::TrimResult tr {0, seq.GetSize()};
+		miint::qc::TrimResult tr {0, static_cast<std::size_t>(seq.GetSize())};
 		const auto seq_bytes = reinterpret_cast<const std::uint8_t *>(seq.GetData());
 		std::size_t best_trim_start = seq.GetSize();
 		for (const auto &cand : *candidates_ptr) {
